+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 Version 0.2.0-dev
 - only first line of sections TEST, ARGS and FILE_EXTERNAL is now parsed
-- sections GET, and COOKIE are parsed now
+- sections GET, COOKIE and PHPDBG are parsed now
 - implemented sections EXPECTREGEX, EXPECTREGEX_EXTERNAL and CAPTURE_STDIO
 
 Version 0.1.0